[#1472] moved record activity
